+++ web/api/src/test/java/org/onosproject/codec/impl/IntentCodecTest.java
@@ -66,8 +66,7 @@ public class IntentCodecTest extends AbstractIntentTest {
 
     private final HostId id1 = hid("12:34:56:78:91:ab/1");
     private final HostId id2 = hid("12:34:56:78:92:ab/1");
-    private final ApplicationId appId =
-            new DefaultApplicationId((short) 3, "test");
+    private final ApplicationId appId = new DefaultApplicationId(3, "test");
     final TrafficSelector emptySelector =
             DefaultTrafficSelector.builder().build();
     final TrafficTreatment emptyTreatment =
